refactored validations folder
