+++ src/lib/db/index.ts
@@ -1,15 +1,6 @@
-import { AdminLevel, PrismaClient, Role } from "@prisma/client";
+import { AdminLevel, PrismaClient } from "@prisma/client";
 
 export const db = new PrismaClient();
-
-export type CompositeUser = {
-  id: string;
-  role: Role | null | undefined;
-} & {
-  name?: string | null | undefined;
-  email?: string | null | undefined;
-  image?: string | null | undefined;
-};
 
 export const adminLevelOrd = {
   [AdminLevel.SUPER]: 3,
